requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.1.1
asgiref=3.8.1=py312hca03da5_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.2.25=hca03da5_0
django=5.1.3=py312hca03da5_0
django-appconf=1.1.0=pypi_0
django-ckeditor=6.7.2=pypi_0
django-compressor=4.5.1=pypi_0
django-js-asset=3.1.2=pypi_0
expat=2.6.4=h313beb8_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
ncurses=6.4=h313beb8_0
openssl=3.0.16=h02f6b3c_0
pillow=11.1.0=pypi_0
pip=25.0=py312hca03da5_0
psycopg2-binary=2.9.10=pypi_0
python=3.12.9=h99e199e_0
rcssmin=1.1.2=pypi_0
readline=8.2=h1a28f6b_0
rjsmin=1.2.2=pypi_0
setuptools=75.8.0=py312hca03da5_0
sqlite=3.45.3=h80987f9_0
sqlparse=0.5.2=py312hca03da5_0
tk=8.6.14=h6ba3021_0
tzdata=2025a=h04d1e81_0
wheel=0.45.1=py312hca03da5_0
xz=5.6.4=h80987f9_1
zlib=1.2.13=h18a0788_1
